fix poc code
